+++ packages/contingency_analysis_pkg/src/toop_engine_contingency_analysis/pandapower/pandapower_helpers/results/node_results.py
@@ -10,7 +10,7 @@
 
 import polars as pl
 from pandapower import pandapowerNet
-from toop_engine_contingency_analysis.pandapower.pandapower_helpers.results.result_constants import (
+from toop_engine_contingency_analysis.pandapower.pandapower_helpers.result_constants import (
     MAX_ALLOWED_VM_DEVIATION,
     ResultConstants,
 )
@@ -18,9 +18,7 @@ from toop_engine_contingency_analysis.pandapower.pandapower_helpers.schemas impo
     PandapowerContingency,
 )
 
-#: Value for numeric result columns on the failed (non-converged) path. NaN, not null: the
-#: pandas pipeline this replaced used ``np.nan`` and downstream polars metrics rely on NaN
-#: semantics (an all-missing ``max()`` must be NaN, not null which collects to None).
+#: Value for numeric result columns on the failed (non-converged) path.
 _MISSING = float("nan")
