+++ packages/contingency_analysis_pkg/tests/pandapower/pandapower_helpers/results/test_get_va_diff_results.py
@@ -16,7 +16,7 @@ from toop_engine_contingency_analysis.pandapower import (
     get_va_diff_results,
 )
 from toop_engine_contingency_analysis.pandapower.pandapower_helpers import VADiffInfo
-from toop_engine_contingency_analysis.pandapower.pandapower_helpers.results.result_constants import ResultConstants
+from toop_engine_contingency_analysis.pandapower.pandapower_helpers.result_constants import ResultConstants
 from toop_engine_contingency_analysis.pandapower.pandapower_helpers.results.switch_results import (
     SwitchElementMappingSchema,
 )
